# Pull Request Description

## Summary
[Provide a brief description of the changes in this PR]

### Issue Reference
Fixes #[Issue Number]

### Motivation and Context
- Why is this change needed?
- What problem does it solve?
- If it fixes an open issue, please link to the issue here

### Dependencies
- List any dependencies that are required for this change
- Include any configuration changes needed
- Note any version updates required

## Type of Change
Please mark the relevant option with an `x`:
- [ ] 🐛 Bug fix (non-breaking change which fixes an issue)
- [ ] ✨ New feature (non-breaking change which adds functionality)
- [ ] 💥 Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 📝 Documentation update (Wiki/README/Code comments)
- [ ] ♻️ Refactor (code improvement without functional changes)
- [ ] 🎨 Style update (formatting, renaming)
- [ ] 🔧 Configuration change
- [ ] 📦 Dependency update

## Testing
- [ ] I have added unit tests that prove my fix is effective or that my
feature works
- [ ] New and existing unit tests pass locally with my changes
- [ ] I have tested this code in the following browsers/environments:
[list environments]

## Quality Checklist
- [ ] I have reviewed my own code before requesting review
- [ ] I have verified there are no other open Pull Requests for the same
update/change
- [ ] All CI/CD pipelines pass without errors or warnings
- [ ] My code follows the established style guidelines of this project
- [ ] I have added necessary documentation (if appropriate)
- [ ] I have commented my code, particularly in complex areas
- [ ] I have made corresponding changes to the README and other relevant
documentation
- [ ] My changes generate no new warnings
- [ ] I have performed a self-review of my own code
- [ ] My code is properly formatted according to project standards

## Screenshots/Recordings (if appropriate)
[Add screenshots or recordings that demonstrate the changes]

## Additional Notes
[Add any additional information that might be helpful for reviewers]

+++ internal/datasources/device_and_app_management/graph_beta/mobile_app/model.go
@@ -38,10 +38,10 @@ type MobileAppModel struct {
 	PublishingState       types.String   `tfsdk:"publishing_state"`
 	IsAssigned            types.Bool     `tfsdk:"is_assigned"`
 	IsFeatured            types.Bool     `tfsdk:"is_featured"`
-	UploadState           types.Int64    `tfsdk:"upload_state"`
-	DependentAppCount     types.Int64    `tfsdk:"dependent_app_count"`
-	SupersededAppCount    types.Int64    `tfsdk:"superseded_app_count"`
-	SupersedingAppCount   types.Int64    `tfsdk:"superseding_app_count"`
+	UploadState           types.Int32    `tfsdk:"upload_state"`
+	DependentAppCount     types.Int32    `tfsdk:"dependent_app_count"`
+	SupersededAppCount    types.Int32    `tfsdk:"superseded_app_count"`
+	SupersedingAppCount   types.Int32    `tfsdk:"superseding_app_count"`
 	RoleScopeTagIds       []types.String `tfsdk:"role_scope_tag_ids"`
 	Categories            []types.String `tfsdk:"categories"`
 }
